OCSADV-265, OCSADV-266: Move presentation logic from recipe to web module for GMOS.
