+++ bundle/edu.gemini.itc.web/src/main/java/edu/gemini/itc/web/servlets/ITCgmosServlet.java
@@ -1,8 +1,11 @@
 package edu.gemini.itc.web.servlets;
 
-import edu.gemini.itc.gmos.GmosRecipe;
+import edu.gemini.itc.shared.GmosParameters;
 import edu.gemini.itc.shared.ITCMultiPartParser;
-import edu.gemini.itc.shared.Recipe;
+import edu.gemini.itc.shared.Parameters;
+import edu.gemini.itc.shared.PlottingDetails;
+import edu.gemini.itc.web.ITCRequest;
+import edu.gemini.itc.web.html.GmosPrinter;
 
 import java.io.PrintWriter;
 
@@ -52,7 +55,10 @@ public final class ITCgmosServlet extends ITCServlet {
      */
     public void writeOutput(final ITCMultiPartParser mpp, final PrintWriter out) {
         out.println("<a href = \"http://www.gemini.edu/sciops/instruments/integration-time-calculators/itc-help\"> Click here for help with the results page.</a>");
-        final Recipe recipe = new GmosRecipe(mpp, out);
-        recipe.writeOutput();
+        final Parameters p  = ITCRequest.parameters(mpp);
+        final GmosParameters ip = ITCRequest.gmosParameters(mpp);
+        final PlottingDetails pdp = ITCRequest.plotParameters(mpp);
+        final GmosPrinter printer = new GmosPrinter(p, ip, pdp, out);
+        printer.writeOutput();
     }
 }
